Remove time zone symlink before ln -s
There's already a symlink pointing to /usr/share/zoneinfo/UTC. It seems
the systemd package is creating that symlink, and the ln -s command fails.

To prevent that, the symlink is deleted before it is created.
To check if the old symlink exists, the rm command is used (instead of
forcing the ln -s command).

When calling a
  pacstrap /mnt systemd
on an empty mountpoint and watching for a /mnt/etc/localtime symlink in
another shell, you can see that the symlink is created when the
post-transcation hooks are running.

I couldn't figure out exactly, but the symlink is created by one of
these hooks:
  :: Running post-transaction hooks...
  (1/3) Updating system user accounts...
  (2/3) Creating temporary files...
  (3/3) Updating udev Hardware Database...